Clarify `*.init` can be used with active segments
On rereading the text of the proposal I *think* this is the intent.
Currently it reads as if `memory.init` and `table.init` can only be used
on passive segments, but the `DataCount` section I believe counts all
data segments, both active and passive, so when verifying function
bodies a segment index isn't known whether to be active or passive.